// pslpwcl/src/app/modules/pslpwcl-common/models/sort-event.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
import { SortDirection } from "./sort-direction";


export interface SortEvent {
  column: string;
  direction: SortDirection;
}